+++ src/createServer.ts
@@ -8,13 +8,13 @@ import passportLocal from "passport-local";
 import * as path from "path";
 import { ssdaPool } from "./db/postgresql_pool";
 import { resolvers } from "./resolvers";
+import { getAuthProvider } from "./util/authProvider";
 import {
   dataFileDataLoader,
   dataRequestDataLoader,
   userDataLoader
 } from "./util/loaders";
 import {
-  getAuthProvider,
   getUserById,
   isAdmin,
   mayViewDataFile,
