+++ tslint.json
@@ -5,6 +5,6 @@
       "severity": "off"
     },
     "no-implicit-dependencies": [true, "dev"],
-    "no-submodule-imports": [false]
+    "no-submodule-imports": [true, "mysql2/node_modules/iconv-lite", "mysql2/promise"]
   }
 }
